run/manager for launching of plugins
