+++ pkg/run/v1/flavor/vanilla/vanilla.go
